M2-4660: moved sticky header from report menu to the separate component to reduce activity list rerender
