+++ src/modules/Dashboard/features/RespondentData/RespondentDataSummary/ReportMenu/ReportMenu.styles.ts
@@ -1,6 +1,6 @@
 import { Box, styled } from '@mui/material';
 
-import { StyledFlexColumn, commonEllipsisStyles, theme, variables } from 'shared/styles';
+import { StyledFlexColumn, theme, variables } from 'shared/styles';
 import { shouldForwardProp } from 'shared/utils';
 
 export const StyledActivity = styled(Box, shouldForwardProp)`
@@ -25,11 +25,6 @@ export const StyledActivity = styled(Box, shouldForwardProp)`
       background-color: ${variables.palette.surface2};
     }
   `}
-`;
-
-export const StyledHeaderContainer = styled(Box, shouldForwardProp)`
-  ${commonEllipsisStyles};
-  ${({ isSticky }: { isSticky: boolean }) => !isSticky && `margin-top: ${theme.spacing(2.4)}`}
 `;
 
 export const StyleContainer = styled(StyledFlexColumn)`
