+++ src/modules/Dashboard/features/RespondentData/RespondentDataSummary/ReportMenu/StickyHeader/StickyHeader.types.ts
@@ -1,0 +1,5 @@
+import { MutableRefObject } from 'react';
+
+export type StickyHeaderProps = {
+  containerRef: MutableRefObject<HTMLElement | null>;
+};
