+++ src/modules/Dashboard/features/RespondentData/RespondentDataSummary/ReportMenu/StickyHeader/index.ts
@@ -1,0 +1,1 @@
+export * from './StickyHeader';
